requirements.txt
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-CLI==0.4.0
Flask-Cors==3.0.7
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.0
idna==2.8
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.15.4
oauth2client==1.5.2
pandas==0.24.2
#pkg-resources==0.0.0
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.3
pytest==4.2.1
python-dateutil==2.8.0
pytz==2018.9
rake-nltk==1.0.4
requests==2.21.0
requests-unixsocket==0.1.5
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.1
selenium==3.141.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
SQLAlchemy==1.3.5
uritemplate==3.0.0
urllib3==1.24.3
Werkzeug==0.15.5
zipp==0.5.2
gunicorn==20.0.4